+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New `uairctl` flag for `listen` subcommand: `-o` or `--override`. Allows to specify a named override created in `uair` config.
 - New `uair` config session propery: `id`. Allows to uniquely identify each session.
 - New `uairctl` subcommand: `jump`. Allows to directly jump to a session with a given id.
+- New `uairctl` subcommand: `reload`. Allows to reload the config file.
 
 #### Removed